[FIX] The same travis error (again).

+++ libcomp/src/TcpConnection.cpp
@@ -370,7 +370,7 @@ void TcpConnection::FlushOutgoingInside(bool closeConnection)
             }
             else
             {
-                mOutgoing.Skip(length);
+                mOutgoing.Skip((uint32_t)length);
 
                 if(0 != mOutgoing.Left())
                 {
